rename thread to agent
